Refactor/structure things so that validator modules can be imported by default (and thus register)
