+++ tests/test_octodns_zone_validators.py
@@ -7,12 +7,12 @@ from unittest import TestCase
 from octodns.record import Record
 from octodns.zone import Zone
 from octodns.zone.exception import ValidationError, ZoneException
+from octodns.zone.mail import MailZoneValidator
 from octodns.zone.validator import (
     ValidationReason,
     ZoneValidator,
     ZoneValidatorRegistry,
 )
-from octodns.zone.validator.mail import MailZoneValidator
 
 
 def _make_zone(name='unit.tests.'):
