+++ tests/test_octodns_zone_validators_mail.py
@@ -6,7 +6,7 @@ from unittest import TestCase
 
 from octodns.record import Record
 from octodns.zone import Zone
-from octodns.zone.validator.mail import MailZoneValidator
+from octodns.zone.mail import MailZoneValidator
 
 
 def _make_zone(name='unit.tests.'):
